% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Bioconductor}
\alias{Bioconductor}
\title{Bioconductor dataset}
\format{An object of class \code{\link{genomic_ressource-class}}.}
\usage{
data(Bioconductor,package="ViSEAGO")
}
\description{
an example of object returned by \code{\link{Bioconductor2GO}} method.
}
\keyword{datasets}
