% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BP_sResults}
\alias{BP_sResults}
\title{BP_sResults dataset}
\format{An object of class \code{\link{enrich_GO_terms-class}}.}
\usage{
data(BP_sResults,package="ViSEAGO")
}
\description{
an example of object returned by \code{\link{merge_enrich_terms}} method.
}
\keyword{datasets}
