% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_cor.R
\name{clusters_cor}
\alias{clusters_cor}
\title{Compute distance matrix between dendrograms partitions.}
\usage{
clusters_cor(clusters, method = "adjusted.rand")
}
\arguments{
\item{clusters}{a \code{list} of \code{\link{GO_clusters-class}} objects,
from \code{\link{GOterms_heatmap}} or \code{\link{GOclusters_heatmap}}, named as \code{character}.}

\item{method}{available methods  ("vi", "nmi", "split.join", "rand", or "adjusted.rand") from \pkg{igraph} package \code{\link[igraph]{compare}} function.}
}
\value{
a distance or a correlation matrix.
}
\description{
Build a distance or correlation matrix between partitions from dendrograms.
}
\examples{
\dontrun{
###################
# clusters to compare
clusters<-base::list(
 Resnik="Resnik_clusters_wardD2",
 Lin="Lin_clusters_wardD2",
 Rel="Rel_clusters_wardD2",
 Jiang="Jiang_clusters_wardD2",
 Wang="Wang_clusters_wardD2"
)

###################
# global dendrogram clustering correlation
clust_cor<-ViSEAGO::clusters_cor(clusters,method="adjusted.rand")
}
}
\references{
Csardi G, Nepusz T: The igraph software package for complex network research, InterJournal, Complex Systems 1695. 2006. http://igraph.org.
}
\seealso{
Other GO_clusters: \code{\link{GO_clusters-class}},
  \code{\link{GOclusters_heatmap}},
  \code{\link{compare_clusters}},
  \code{\link{show_heatmap}}, \code{\link{show_table}}
}
\concept{GO_clusters}
