% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_clusters.R
\name{compare_clusters}
\alias{compare_clusters}
\title{Heatmap to compare partitions}
\usage{
compare_clusters(clusters)
}
\arguments{
\item{clusters}{a \code{list} of named \code{\link{GO_clusters-class}} objects,
from \code{\link{GOterms_heatmap}} or \code{\link{GOclusters_heatmap}} methods.}
}
\value{
an interactive javascript heatmap.
}
\description{
Build an interactive heatmap of the common GO terms frequency between several partitions.
}
\details{
Build an interactive heatmap of common GO terms frequency between partitions from several
\code{\link{GO_clusters-class}} objects.
}
\examples{
\dontrun{
###################
# clusters to compare
clusters<-base::list(
 Resnik="Resnik_clusters_wardD2",
 Lin="Lin_clusters_wardD2",
 Rel="Rel_clusters_wardD2",
 Jiang="Jiang_clusters_wardD2",
 Wang="Wang_clusters_wardD2"
)

###################
# clusters content comparisons
clusters_comp<-ViSEAGO::compare_clusters(clusters)
}
}
\references{
Carson Sievert, Chris Parmer, Toby Hocking, Scott Chamberlain, Karthik Ram, Marianne Corvellec and Pedro Despouy (2017).
 plotly: Create Interactive Web Graphics via 'plotly.js'. R package version 4.6.0. https://CRAN.R-project.org/package=plotly
}
\seealso{
Other GO_clusters: \code{\link{GO_clusters-class}},
  \code{\link{GOclusters_heatmap}},
  \code{\link{clusters_cor}}, \code{\link{show_heatmap}},
  \code{\link{show_table}}
}
\concept{GO_clusters}
