% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntrezGene_orthologs.R
\name{EntrezGene_orthologs}
\alias{EntrezGene_orthologs}
\title{Download EntrezGene orthologs groups.}
\usage{
EntrezGene_orthologs()
}
\value{
a \code{\link[data.table]{data.table}}.
}
\description{
Download EntrezGene orthologs groups.
}
\details{
Internal function used by \code{\link{annotate}} in order to download orthologs_groups from NCBI Annotation pipeline
 stored in the \href{ftp://ftp.ncbi.nih.gov/gene/DATA/gene_orthologs.gz}{gene_group.gz} file.
}
\examples{
###################
# Organism taxid, Scientific name and common name
ViSEAGO::EntrezGene_orthologs()
}
\keyword{internal}
