% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\docType{class}
\name{fgsea-class}
\alias{fgsea-class}
\title{fgsea class object definition.}
\description{
This class is invoked by \code{\link{runfgsea}} method in order to store results.
}
\section{Slots}{

\describe{
\item{\code{description}}{a \code{character} string with database source, date of stamp, and target species GO annotation.}

\item{\code{method}}{fgsea method used.}

\item{\code{params}}{a \code{list} containing used input parameters for perform \code{\link[fgsea]{fgseaSimple}} or \code{\link[fgsea]{fgseaMultilevel}}.}

\item{\code{input}}{a \code{list} containing input values.}

\item{\code{data}}{a \code{list} containing \code{data.table} fgsea procedure output.}
}}

