% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2GO.R
\docType{class}
\name{gene2GO-class}
\alias{gene2GO-class}
\title{gene2GO class object definition.}
\description{
This class is invoked by \code{\link{annotate}} method in order to store GO annotations for each category (MF, BP, CC).
}
\section{Slots}{

\describe{
\item{\code{db}}{database source in \code{character}.}

\item{\code{stamp}}{date of stamp  in \code{character}.}

\item{\code{organism}}{target species GO annotation  in \code{character}.}

\item{\code{MF}}{a \code{list} containing  GO terms for Molecular Function (MF) category for each gene element.}

\item{\code{BP}}{a \code{list} containing  GO terms for Biological Process (BP) category for each gene element.}

\item{\code{CC}}{a \code{list} containing  GO terms for Cellular Component (CC) category for each gene element.}
}}

\seealso{
Other GO_terms: \code{\link{GOcount}},
  \code{\link{GOterms_heatmap}}, \code{\link{annotate}},
  \code{\link{create_topGOdata}},
  \code{\link{merge_enrich_terms}}
}
\concept{GO_terms}
