% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_ressource.R
\docType{class}
\name{genomic_ressource-class}
\alias{genomic_ressource-class}
\title{genomic_ressource class object definition.}
\description{
This class stores the annotations and associated metadata obtained by \code{\link{Bioconductor2GO}},
\code{\link{EntrezGene2GO}}, \code{\link{Ensembl2GO}}, or \code{\link{Uniprot2GO}} methods.
}
\section{Slots}{

\describe{
\item{\code{db}}{name of database used (Bioconductor, EntrezGene, Ensembl, or Uniprot).}

\item{\code{stamp}}{date of stamp (for Bioconductor, EntrezGene, and Uniprot), or annotation version for Ensembl database.}

\item{\code{data}}{GO annotations from \code{\link{EntrezGene2GO}} method.}

\item{\code{organisms}}{informations about species/datasets availables.}

\item{\code{mart}}{Ensembl mart from \code{\link{Ensembl2GO}} method.}
}}

\seealso{
Other genomic_ressource: \code{\link{Bioconductor2GO}},
  \code{\link{Ensembl2GO}}, \code{\link{EntrezGene2GO}},
  \code{\link{Uniprot2GO}}, \code{\link{annotate}},
  \code{\link{available_organisms}}, \code{\link{taxonomy}}
}
\concept{genomic_ressource}
