% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_clusters.R
\docType{class}
\name{GO_clusters-class}
\alias{GO_clusters-class}
\title{GO_clusters class object}
\description{
This class is invoked by \code{\link{GOterms_heatmap}} and  \code{\link{GOclusters_heatmap}} methods to store all results produced.
}
\section{Slots}{

\describe{
\item{\code{ont}}{ontology used "MF", "BP", or "CC".}

\item{\code{enrich_GOs}}{\code{\link{enrich_GO_terms-class}} object.}

\item{\code{IC}}{Information Content (IC).}

\item{\code{terms_dist}}{distance between GO terms based on semantic similiarity.}

\item{\code{clusters_dist}}{distance between GO groups based on semantic similiarity.}

\item{\code{hcl_params}}{Hierarchical clustering parameters used.}

\item{\code{dendrograms}}{GO terms and samples \code{dendrograms}.}

\item{\code{samples.gp}}{samples groups.}

\item{\code{heatmap}}{GO terms and GO groups heatmaps.}
}}

\seealso{
Other GO_clusters: 
\code{\link{GOclusters_heatmap}()},
\code{\link{clusters_cor}()},
\code{\link{compare_clusters}()},
\code{\link{show_heatmap}()},
\code{\link{show_table}()}
}
\concept{GO_clusters}
