% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_SS.R
\docType{class}
\name{GO_SS-class}
\alias{GO_SS-class}
\title{GO_SS class object definition.}
\description{
This class is invoked by \code{\link{build_GO_SS}} method in order to store \code{\link{enrich_GO_terms-class}} object, Information Content (IC),
and GO terms or groups distances objects based on semantic similarity.
}
\section{Slots}{

\describe{
\item{\code{db}}{should be "Bioconductor", "EntrezGene", "Ensembl", or "Uniprot" ressource name.}

\item{\code{stamp}}{date of stamp or annotation version if available in \code{character}.}

\item{\code{organism}}{target species GO anotation in \code{character}.}

\item{\code{ont}}{used ontology with "MF", "BP", or "CC".}

\item{\code{topGO}}{\code{list} with topGO objects summary informations.}

\item{\code{IC}}{Information Content (IC)}

\item{\code{enrich_GOs}}{\code{\link{merge_enrich_terms}} output object (\code{\link{enrich_GO_terms-class}} object).}

\item{\code{terms_dist}}{\code{list} of GO terms or groups distances objects based on semantic similarity.}
}}

\seealso{
Other GO_semantic_similarity: \code{\link{build_GO_SS}},
  \code{\link{compute_SS_distances}}
}
\concept{GO_semantic_similarity}
