% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{myGENE2GO}
\alias{myGENE2GO}
\title{myGENE2GO dataset}
\format{An object of class \code{\link{gene2GO-class}}.}
\usage{
data(myGENE2GO,package="ViSEAGO")
}
\description{
an example of object returned by \code{\link{annotate}} method.
}
\keyword{datasets}
