% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{myGOs}
\alias{myGOs}
\title{myGOs dataset}
\format{An object of class \code{\link{GO_SS-class}}.}
\usage{
data(myGOs,package="ViSEAGO")
}
\description{
an example of object returned by \code{\link{build_GO_SS}} method
from mouse functional analysis of mouse mammary gland RNA-Seq (2_mouse_bioconductor vignette)
}
\keyword{datasets}
