% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapper.R
\name{overLapper}
\alias{overLapper}
\title{build all sets combinations intersections}
\usage{
overLapper(setlist)
}
\value{
a \code{\link[base]{list}}.
}
\description{
build all sets combinations intersections
}
\details{
This internal function use build all sets combinations intersections needed for \code{\link{upset}}.
}
\examples{
\dontrun{
# build all intersections combinations
ViSEAGO::overLapper(setlist)
}
}
\seealso{
Other enrich_GO_terms: 
\code{\link{Upset}()},
\code{\link{enrich_GO_terms-class}},
\code{\link{show_heatmap}()},
\code{\link{show_table}()}

Other visualization: 
\code{\link{GOclusters_heatmap}()},
\code{\link{GOcount}()},
\code{\link{GOterms_heatmap}()},
\code{\link{Upset}()},
\code{\link{available_organisms}()},
\code{\link{show_heatmap}()},
\code{\link{show_table}()}
}
\concept{enrich_GO_terms}
\concept{visualization}
\keyword{internal}
