% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdiagram.R
\name{pkgdiagram}
\alias{pkgdiagram}
\title{Display the ViSEAGO package diagram.}
\usage{
pkgdiagram(x = c("white", "white", "white", "white", "white"))
}
\arguments{
\item{x}{rectangles colors for the five subgraphs with all white by defaults}
}
\value{
an html diagram
}
\description{
This function build the ViSEAGO package diagram available displayed in the vignette.
}
\details{
This function use \pkg{DiagrammeR} package \code{\link[DiagrammeR]{grViz}} to display the ViSEAGO methods diagram with colored or not subgraphs rectangles.
}
\examples{
###################
# print diagram without focus
ViSEAGO::pkgdiagram(x=c("white","white","white","white","white"))

###################
# print diagram with focus one the gene list
ViSEAGO::pkgdiagram(x=c("black","white","white","white","white"))

###################
# print diagram with a focus on the annotation step
ViSEAGO::pkgdiagram(x=c("white","black","white","white","white"))

###################
# print diagram with a focus on the enrichment step
ViSEAGO::pkgdiagram(x=c("white","white","black","white","white"))

###################
# print diagram with a focus on the Semantic Similarity step
ViSEAGO::pkgdiagram(x=c("white","white","white","black","white"))

###################
 # print diagram with focus on the visualization step
ViSEAGO::pkgdiagram(x=c("white","white","white","white","black"))
}
\references{
Knut Sveidqvist, Mike Bostock, Chris Pettitt, Mike Daines, Andrei Kashcha and Richard Iannone (2017). DiagrammeR: Create Graph Diagrams and Flowcharts Using R.
R package version 0.9.0. https://CRAN.R-project.org/package=DiagrammeR.
}
\keyword{internal}
