% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_heatmap.R
\name{show_heatmap}
\alias{show_heatmap}
\title{Display an interactive or static heatmap.}
\usage{
show_heatmap(object, type, file = NULL)
}
\arguments{
\item{object}{a \code{\link{GO_clusters-class}} object from \code{\link{GOterms_heatmap}} or \code{\link{GOclusters_heatmap}}.}

\item{type}{could be "GOterms" to display GOterms clustering heatmap, or "GOclusters" to display GOclusters heatmap.}

\item{file}{static image output file name (default to NULL).}
}
\description{
Display a heatmap in interactive or static mode.
}
\details{
This method displays an interactive heatmap (if \code{file}=NULL) from \code{\link{GO_clusters-class}} object for "GOterms" or "GOclusters" type.\cr
A static png image could be printed by setting \code{file} argument.
}
\examples{
###################
# load object
utils::data(
 Wang_clusters_wardD2,
 package="ViSEAGO"
)

##################
# Display GO terms heatmap
ViSEAGO::show_heatmap(Wang_clusters_wardD2,"GOterms")

##################
# Display GO clusters heatmap
ViSEAGO::show_heatmap(Wang_clusters_wardD2,"GOclusters")

##################
# Print GO terms heatmap
ViSEAGO::show_heatmap(Wang_clusters_wardD2,"GOterms","./data/output/GOterms_heatmap.png")

##################
# Print GO clusters heatmap
ViSEAGO::show_heatmap(Wang_clusters_wardD2,"GOclusters","./data/output/GOclusters_heatmap.png")
}
\seealso{
Other enrich_GO_terms: \code{\link{Upset}},
  \code{\link{enrich_GO_terms-class}},
  \code{\link{overLapper}}, \code{\link{show_table}}

Other GO_clusters: \code{\link{GO_clusters-class}},
  \code{\link{GOclusters_heatmap}},
  \code{\link{clusters_cor}},
  \code{\link{compare_clusters}}, \code{\link{show_table}}

Other visualization: \code{\link{GOclusters_heatmap}},
  \code{\link{GOcount}}, \code{\link{GOterms_heatmap}},
  \code{\link{Upset}}, \code{\link{available_organisms}},
  \code{\link{overLapper}}, \code{\link{show_table}}
}
\concept{GO_clusters}
\concept{enrich_GO_terms}
\concept{visualization}
