% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\title{Display Organism Scientific and common name from taxid.}
\usage{
taxonomy(...)
}
\value{
a \code{\link[data.table]{data.table}}.
}
\description{
Display Organism Scientific and common name from taxid.
}
\details{
This internal function use \href{https://www.ncbi.nlm.nih.gov/books/NBK25500/}{E-utilities} summary to display in
 \code{\link[data.table]{data.table}} Organism Scientific name and common name, form a \code{\link[base]{vector}} of taxid.
}
\examples{
\dontrun{
###################
# Organism Scientific and common name from taxid
ViSEAGO::taxonomy(taxid="9031")
}
}
\seealso{
Other genomic_ressource: \code{\link{Bioconductor2GO}},
  \code{\link{Ensembl2GO}}, \code{\link{EntrezGene2GO}},
  \code{\link{Uniprot2GO}}, \code{\link{annotate}},
  \code{\link{available_organisms}},
  \code{\link{genomic_ressource-class}}
}
\concept{genomic_ressource}
\keyword{internal}
