% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Wang_clusters_wardD2}
\alias{Wang_clusters_wardD2}
\title{Wang_clusters_wardD2 dataset}
\format{An object of class \code{\link{GO_clusters-class}}.}
\usage{
data(Wang_clusters_wardD2,package="ViSEAGO")
}
\description{
an example of object returned by \code{\link{GOclusters_heatmap}} method.
}
\keyword{datasets}
