% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_GO_terms.R
\docType{class}
\name{enrich_GO_terms-class}
\alias{enrich_GO_terms-class}
\title{enrich_GO_terms class object definition.}
\description{
This class is invoked by \code{\link{merge_enrich_terms}} method in order to store the merged \code{data.table} and associated metadata.
}
\section{Slots}{

\describe{
\item{\code{same_genes_background}}{logical.}

\item{\code{input}}{a list containing named elements. Each element must contain the name of \code{\link[topGO]{topGOdata-class}}
object created by \code{\link{create_topGOdata}} method and the associated  \code{\link[topGO]{topGOresult-class}}
object(s) to combinate (see examples in \code{\link{merge_enrich_terms}}).}

\item{\code{ont}}{ontology used "MF", "BP", or "CC".}

\item{\code{topGO}}{a \code{list} with topGO objects summary informations.}

\item{\code{data}}{a merged \code{data.table}  of enriched GO terms (p<0.01) in at least once with GO descriptions and statistical values.}
}}

\seealso{
Other enrich_GO_terms: \code{\link{Upset}},
  \code{\link{overLapper}}, \code{\link{show_heatmap}},
  \code{\link{show_table}}
}
\concept{enrich_GO_terms}
