% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSplot.R
\name{MDSplot}
\alias{MDSplot}
\title{Multi Dimensional Scale (MDS) plot}
\usage{
MDSplot(object, show_clusters = FALSE, file = NULL)
}
\arguments{
\item{object}{a \code{\link{GO_SS-class}} or \code{\link{GO_clusters-class}} objects from distances computed with \code{\link{compute_SS_distances}}.}

\item{show_clusters}{boolean (default to FALSE)}

\item{file}{static image output file name (default to NULL).}
}
\value{
a MDS plot.
}
\description{
Generate a Multi Dimensional Scale (MDS) plot from distance objects.
}
\details{
This method build and display the javascript MDSplot (if \code{file}=NULL) from \code{\link{GO_SS-class}} or \code{\link{GO_clusters-class}}
objects.\cr
A static png image could be printed by setting \code{file} argument.
}
\examples{
\dontrun{
###################
# build MDS plot for a GO_SS-class distance object
ViSEAGO::MDSplot(myGOs)

###################
# build MDS plot for a GO_clusters-class distance object, highlighting GO terms clusters.
ViSEAGO::MDSplot(Wang_clusters_wardD2)

###################
# build MDS plot for a GO_clusters-class distance object, highlighting GO groups clusters.
ViSEAGO::MDSplot(
 Wang_clusters_wardD2,
 show_clusters=TRUE
)
}
}
\concept{GO_terms GO_clusters semantic_similarity visualization}
